requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn
#pystan==2.19.1.1
prophet

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs
uvicorn
fastapi

# utilities
six>=1.14
joblib
memoized-property
termcolor

# Web
streamlit
streamlit_option_menu
# streamlit.components.v1
streamlit_folium
# folium
# common
# os
plost
